+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- We now add all metadata (including ground truth labels, if applicable) to the model
+  cache when debug mode is enabled (with `--debug` or `debug=True`). We have added a
+  [section in the documentation](missing) on how to use this feature.
+
 ### Fixed
 
 - v16.15.0 introduced an error related to the parsing of safetensors metadata from